Reorganized the components folder by creating a subfolder for each component, and moving the codes to their respective subdirectories.
