+++ src/components/Header/Header.css
@@ -6,7 +6,7 @@ header {
 
 header .bg-image {
   position: absolute;
-  background: #333 url("../images/internships/thread-internship-1month.JPG");
+  background: #333 url("../../images/internships/thread-internship-1month.JPG");
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
